Draw arcs as lines, with paired ends as rectangles
The ellipseband! function was removed and replaced with drawloop!
function, which draws a single line (instead of a thick arc) from the
midpoint of each paired end. The paired ends have rectangles to indicate
the width of each end. The plotloops! function's keyword arguments have
been updated accordingly. Additionally, the length and step keywords for
plotloops! have been removed and replaced with resolution, which will
take in a number of points to draw across the provided range on the
x-axis. The loops will be drawn to a consistent resolution by that
keyword argument. Tests and documentation have been updated, and the
example figure in the README has been removed.