Rename and edit liftover functions, documentation
`liftover_gwas!` is now `liftover_sumstats` to be in line with
`mungesumstats!`. Additionally, `liftover_gwas!` can now be applied to
AbstractVector{<:AbstractDataFrame}, in line with `mungesumstats`
behavior. `parsechain` drops the "chr" prefix to be in line with
`mungesumstats!` behavior. Function parameter `echain` renamed to
`chain` (originally named `echain` for "expanded chain", but was
needlessly verbose since we never use the output from `parsechain`).
Documentation for the liftover functions added to the end of the summary
statistics tutorial and mentioned in the GENCODE GTF parsing tutorial.
Tests updated to test liftover functionality over a vector.